# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gachita API",
        "title": "Gachita API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/rooms": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "내 방 목록",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/RoomResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "방 생성",
                "parameters": [
                    {
                        "description": "방 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/RoomResponse"
                        }
                    }
                }
            }
        },
        "/api/rooms/join": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "방 입장",
                "parameters": [
                    {
                        "description": "초대 코드",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/JoinRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/JoinRoomResponse"
                        }
                    }
                }
            }
        },
        "/api/rooms/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "room"
                ],
                "summary": "방 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RoomDetailResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateRoomRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "학교 카풀"
                },
                "openchat_url": {
                    "type": "string",
                    "x-nullable": true,
                    "example": "https://open.kakao.com/o/xxxxx"
                }
            }
        },
...
        "JoinRoomRequest": {
            "type": "object",
            "required": [
                "invite_code"
            ],
            "properties": {
                "invite_code": {
                    "type": "string",
                    "example": "a1b2c3d4"
                }
            }
        },
        "JoinRoomResponse": {
            "type": "object",
            "required": [
                "id",
                "invite_code",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "invite_code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "RoomDetailResponse": {
            "type": "object",
            "required": [
                "id",
                "invite_code",
                "members",
                "name",
                "owner_id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "invite_code": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RoomMemberResponse"
                    }
                },
                "name": {
                    "type": "string"
                },
                "openchat_url": {
                    "type": "string",
                    "x-nullable": true
                },
                "owner_id": {
                    "type": "string"
                }
            }
        },
        "RoomMemberResponse": {
            "type": "object",
            "required": [
                "email",
                "nickname",
                "user_id"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "RoomResponse": {
            "type": "object",
            "required": [
                "id",
                "invite_code",
                "name",
                "owner_id"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "invite_code": {
                    "type": "string",
                    "example": "a1b2c3d4"
                },
                "name": {
                    "type": "string",
                    "example": "학교 카풀"
                },
                "openchat_url": {
                    "type": "string",
                    "x-nullable": true
                },
                "owner_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}